+++ frontend/src/pages/Support.js
@@ -2,11 +2,12 @@ import React from "react";
 // Connecting the NavBar to the pages
 import Navbar from "../components/NavBar";
 
-export default function Community() {
+export default function Support() {
   return (
-    <><Navbar /><div>
-      <h1>Community page</h1>
+    <><Navbar /><div> 
+      <h1>Support</h1>
     </div></>
+
   );
 }